Move ADK mocks to conftest.py for global setup
- Moved comprehensive ADK mock hierarchy to conftest.py for global setup
- Ensures all tests use the same MockLlmAgent base class
- Prevents mock conflicts when test files are loaded in different orders
- Removed duplicate mock setup from test_smartfix_llm_agent.py and test_agent_infrastructure.py

This fixes test_get_extended_model_with_reference when running via make test.

Test suite now down to 1 failure (unrelated test_push_branch).
